Workaround the depfile .d handling
Bazel's code is a little finicky, the checking can be disabled with
--features=-parse_showincludes. But then builds will fail because the .d
file is a not mandatory declared output file. And msvc does not create
it. The code path to parse it will handle its absence, but when we skip
the parsing the internal workaround is also bypassed. So the solution is
to create the file, and then not parse it.